[#52] refactor: shop domain dto 수정
