+++ src/main/java/me/myeats/delivery/shop/controller/MenuController.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import me.myeats.delivery.common.jwt.owner.CurrentOwner;
 import me.myeats.delivery.owner.domain.Owner;
-import me.myeats.delivery.shop.dto.MenuSaveRequestDto;
+import me.myeats.delivery.shop.dto.request.MenuSaveRequestDto;
 import me.myeats.delivery.shop.service.MenuService;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
