+++ src/main/java/me/myeats/delivery/shop/controller/ShopController.java
@@ -3,10 +3,14 @@ package me.myeats.delivery.shop.controller;
 import lombok.RequiredArgsConstructor;
 import me.myeats.delivery.common.jwt.owner.CurrentOwner;
 import me.myeats.delivery.owner.domain.Owner;
-import me.myeats.delivery.shop.dto.ShopSaveRequestDto;
-import me.myeats.delivery.shop.dto.ShopSearchResponseDto;
+import me.myeats.delivery.shop.dto.request.ShopSaveRequestDto;
+import me.myeats.delivery.shop.dto.response.ShopSearchResponseDto;
 import me.myeats.delivery.shop.service.ShopService;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.Valid;
