+++ src/main/java/me/myeats/delivery/shop/dto/OptionGroupSpecDto.java
@@ -1,13 +1,14 @@
 package me.myeats.delivery.shop.dto;
 
+import lombok.AccessLevel;
 import lombok.Builder;
-import lombok.Data;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.util.List;
 
-@Data
-@NoArgsConstructor
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class OptionGroupSpecDto {
 
     private String name;
