+++ src/main/java/me/myeats/delivery/shop/dto/OptionSpecDto.java
@@ -1,11 +1,12 @@
 package me.myeats.delivery.shop.dto;
 
+import lombok.AccessLevel;
 import lombok.Builder;
-import lombok.Data;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Data
-@NoArgsConstructor
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class OptionSpecDto {
 
     private String name;
