+++ src/main/java/me/myeats/delivery/shop/dto/request/MenuSaveRequestDto.java
@@ -1,13 +1,15 @@
-package me.myeats.delivery.shop.dto;
+package me.myeats.delivery.shop.dto.request;
 
+import lombok.AccessLevel;
 import lombok.Builder;
-import lombok.Data;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
+import me.myeats.delivery.shop.dto.MenuDto;
 
 import java.util.List;
 
-@Data
-@NoArgsConstructor
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class MenuSaveRequestDto {
 
     List<MenuDto> menus;
