+++ src/main/java/me/myeats/delivery/shop/dto/response/ShopSearchResponseDto.java
@@ -1,11 +1,15 @@
-package me.myeats.delivery.shop.dto;
+package me.myeats.delivery.shop.dto.response;
 
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
+import me.myeats.delivery.shop.dto.ShopDto;
 
 import java.util.List;
 
 @Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class ShopSearchResponseDto {
 
     private Integer size;
