+++ src/test/java/me/myeats/delivery/shop/controller/MenuControllerTest.java
@@ -8,9 +8,9 @@ import me.myeats.delivery.shop.domain.MenuRepository;
 import me.myeats.delivery.shop.domain.Shop;
 import me.myeats.delivery.shop.domain.ShopRepository;
 import me.myeats.delivery.shop.dto.MenuDto;
-import me.myeats.delivery.shop.dto.MenuSaveRequestDto;
 import me.myeats.delivery.shop.dto.OptionGroupSpecDto;
 import me.myeats.delivery.shop.dto.OptionSpecDto;
+import me.myeats.delivery.shop.dto.request.MenuSaveRequestDto;
 import me.myeats.delivery.test.fixture.OwnerFixtures;
 import me.myeats.delivery.test.fixture.ShopFixtures;
 import me.myeats.delivery.test.security.WithCustomOwner;
