+++ src/test/java/me/myeats/delivery/shop/service/MenuServiceTest.java
@@ -6,9 +6,9 @@ import me.myeats.delivery.shop.domain.MenuRepository;
 import me.myeats.delivery.shop.domain.Shop;
 import me.myeats.delivery.shop.domain.ShopRepository;
 import me.myeats.delivery.shop.dto.MenuDto;
-import me.myeats.delivery.shop.dto.MenuSaveRequestDto;
 import me.myeats.delivery.shop.dto.OptionGroupSpecDto;
 import me.myeats.delivery.shop.dto.OptionSpecDto;
+import me.myeats.delivery.shop.dto.request.MenuSaveRequestDto;
 import me.myeats.delivery.test.fixture.ShopFixtures;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
