+++ src/test/java/me/myeats/delivery/shop/service/ShopServiceTest.java
@@ -4,8 +4,8 @@ import me.myeats.delivery.common.money.Money;
 import me.myeats.delivery.owner.domain.OwnerRepository;
 import me.myeats.delivery.shop.domain.Shop;
 import me.myeats.delivery.shop.domain.ShopRepository;
-import me.myeats.delivery.shop.dto.ShopSaveRequestDto;
-import me.myeats.delivery.shop.dto.ShopSearchResponseDto;
+import me.myeats.delivery.shop.dto.request.ShopSaveRequestDto;
+import me.myeats.delivery.shop.dto.response.ShopSearchResponseDto;
 import me.myeats.delivery.test.fixture.OwnerFixtures;
 import me.myeats.delivery.test.fixture.ShopFixtures;
 import org.junit.jupiter.api.BeforeEach;
